Adding auth index page and fixing back buttons
