+++ src/navigators/navigator.js
@@ -3,14 +3,16 @@ import AppStack from './appNavigator'
 import LogInScreen from '../Auth/LogIn'
 import SignUpScreen from '../Auth/SignUp'
 import AuthLoadingScreen from '../Auth/AuthLoadingScreen'
+import IndexScreen from '../Auth/index'
 
 const AuthStack = createStackNavigator(
   {
+    AuthRoot: IndexScreen,
     SignUp: SignUpScreen,
     LogIn: LogInScreen,
   },
   {
-    initialRouteName: 'SignUp',
+    initialRouteName: 'AuthRoot',
   }
 );
